backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
openai==1.54.0  # For OpenAI Agent SDK
google-generativeai==0.8.0  # For Gemini direct API
qdrant-client==1.12.0
# asyncpg==0.29.0  # Optional: Only needed for US4 (Persistence, P3). Requires C++ Build Tools on Windows
python-multipart==0.0.12
tiktoken==0.8.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
